refactor: Move DateContext and some components to DateContainer
