+++ src/HomeView/index.tsx
@@ -1,0 +1,20 @@
+import { CurrentDateBox, DateList } from '../DateContainer'
+import useTimer from '../shared/hooks/useTimer'
+
+const HomeView = () => {
+  const [date] = useTimer()
+
+  return (
+    <div>
+      <CurrentDateBox
+        date={date}
+      />
+
+      <DateList
+        date={date}
+      />
+    </div>
+  )
+}
+
+export default HomeView
