--- src/common/contexts/DateContext/index.ts
@@ -1,2 +1,0 @@
-export { DateProvider } from "./DateProvider"
-export { useDateContext } from "./useDateContext"
