+++ src/shared/hooks/useTimer.ts
